KOGITO-3450 Remove read only option
